Rename active to event frames summary.
