+++ app/dashes/eventFramesSummary/layout.py
