+++ app/templates/base.html
@@ -58,12 +58,12 @@
                     </li>
                     <li class = "dropdown"><a href = "#" class = "dropdown-toggle" data-toggle = "dropdown">Dashes <span class = "caret"></span></a>
                         <ul class = "dropdown-menu">
-                            <li><a href = '{{ url_for("dashes.activeEventFramesSummary") }}'>Active Event Frames Summary</a></li>
                             <li><a href = '{{ url_for("dashes.elementsGraph") }}'>Elements Graph</a></li>
                             <li><a href = '{{ url_for("dashes.elementsSummary") }}'>Elements Summary</a></li>
                             <li><a href = '{{ url_for("dashes.eventFrameGraph") }}'>Event Frame Graph</a></li>
                             <li><a href = '{{ url_for("dashes.eventFrameGroupSummary") }}'>Event Frame Group Summary</a></li>
                             <li><a href = '{{ url_for("eventFramesOverlay.selectEventFrameTemplate") }}'>Event Frames Overlay</a></li>
+                            <li><a href = '{{ url_for("dashes.eventFramesSummary") }}'>Event Frames Summary</a></li>
                             <li><a href = '{{ url_for("dashes.tagValuesGraph") }}'>Tag Values Graph</a></li>
                         </ul>
                     </li>
